+++ server/src/alerts/target.rs
@@ -16,24 +16,122 @@
  *
  */
 
+use std::{
+    sync::{Arc, Mutex},
+    time::Duration,
+};
+
 use serde::{Deserialize, Serialize};
 
 use super::{AlertState, CallableTarget, Context};
 
+#[derive(Debug, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct Target {
+    #[serde(flatten)]
+    pub target: TargetType,
+    #[serde(flatten)]
+    pub timeout: Option<Timeout>,
+}
+
+impl Target {
+    pub fn call(&self, context: Context) {
+        if let Some(ref timeout) = self.timeout {
+            let resolves = context.alert_state;
+            let mut state = timeout.state.lock().unwrap();
+
+            match resolves {
+                AlertState::SetToFiring => {
+                    state.alert_state = AlertState::Firing;
+
+                    if !state.timed_out {
+                        // set state
+                        state.timed_out = true;
+                        state.awaiting_resolve = true;
+                        drop(state);
+
+                        self.spawn_timeout_task(timeout, context.clone());
+                        call_target(self.target.clone(), context)
+                    }
+                }
+                AlertState::Resolved => {
+                    state.alert_state = AlertState::Listening;
+                    if state.timed_out {
+                        // if in timeout and resolve came in, only process if it's the first one ( awaiting resolve )
+                        if state.awaiting_resolve {
+                            state.awaiting_resolve = false;
+                        } else {
+                            // no further resolve will be considered in timeout period
+                            return;
+                        }
+                    }
+
+                    call_target(self.target.clone(), context);
+                }
+                _ => unreachable!(),
+            }
+        } else {
+            // Without timeout there is no alert state to consider other than the one returned on `resolves`
+            call_target(self.target.clone(), context);
+        }
+    }
+
+    fn spawn_timeout_task(&self, timeout: &Timeout, alert_context: Context) {
+        let state = Arc::clone(&timeout.state);
+        let timeout = timeout.timeout;
+        let target = self.target.clone();
+
+        actix_web::rt::spawn(async move {
+            const RETRIES: usize = 10;
+            // sleep for timeout period
+            for _ in 0..RETRIES {
+                tokio::time::sleep(timeout).await;
+                let mut state = state.lock().unwrap();
+                if state.alert_state == AlertState::Firing {
+                    // it is still firing .. sleep more and come back
+                    state.awaiting_resolve = true;
+
+                    call_target(target.clone(), alert_context.clone())
+                } else {
+                    state.timed_out = false;
+                    return;
+                }
+            }
+
+            // fallback for if this task only observed FIRING on all RETRIES
+            // Stream might be dead and sending too many alerts is not great
+            // Send and alert stating that this alert will only work once it has seen a RESOLVE
+            state.lock().unwrap().timed_out = false;
+            let mut context = alert_context;
+            context.message = format!(
+                "Triggering alert did not resolve itself after {} retries, This alert is paused until it resolves", RETRIES
+            );
+            // Send and exit this task.
+            call_target(target, context);
+        });
+    }
+}
+
+fn call_target(target: TargetType, context: Context) {
+    actix_web::rt::spawn(async move {
+        target.call(&context);
+    });
+}
+
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 #[serde(tag = "type")]
-pub enum Target {
+pub enum TargetType {
     Slack(SlackWebHook),
     #[serde(alias = "webhook")]
     Other(OtherWebHook),
 }
 
-impl Target {
-    pub fn call(&self, payload: impl AsRef<Context>) {
+impl TargetType {
+    pub fn call(&self, payload: &Context) {
         match self {
-            Target::Slack(target) => target.call(payload.as_ref()),
-            Target::Other(target) => target.call(payload.as_ref()),
+            TargetType::Slack(target) => target.call(payload),
+            TargetType::Other(target) => target.call(payload),
         }
     }
 }
@@ -101,4 +199,19 @@ impl CallableTarget for OtherWebHook {
             log::error!("Couldn't make call to webhook, error: {}", e)
         }
     }
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct Timeout {
+    #[serde(with = "humantime_serde")]
+    pub timeout: Duration,
+    #[serde(skip)]
+    pub state: Arc<Mutex<TimeoutState>>,
+}
+
+#[derive(Debug, Default, Clone, Copy)]
+pub struct TimeoutState {
+    pub alert_state: AlertState,
+    pub timed_out: bool,
+    pub awaiting_resolve: bool,
 }
